+++ user.js
@@ -28,7 +28,6 @@ user_pref("user.theme.brave", false);
 user_pref("user.theme.ayu", false);
 user_pref("user.theme.rose-pine", false);
 user_pref("user.theme.pywalfox", false);
-user_pref("user.theme.matugen", false);
 user_pref("user.theme.xtras.color-scheme.force.consistency", false);
 
 /* nav bar */